feat(language-client): separate workspaceSymbol
