+++ src/language-client/index.ts
@@ -3,7 +3,7 @@
 import cp from 'child_process'
 import fs from 'fs'
 import path from 'path'
-import { createClientPipeTransport, createClientSocketTransport, Disposable, DocumentFilter, generateRandomPipeName, IPCMessageReader, IPCMessageWriter, StreamMessageReader, StreamMessageWriter } from 'vscode-languageserver-protocol/node'
+import { createClientPipeTransport, createClientSocketTransport, Disposable, generateRandomPipeName, IPCMessageReader, IPCMessageWriter, StreamMessageReader, StreamMessageWriter } from 'vscode-languageserver-protocol/node'
 import { ServiceStat } from '../types'
 import { disposeAll } from '../util'
 import * as Is from '../util/is'
@@ -23,6 +23,7 @@ import ChildProcess = cp.ChildProcess
 import { CallHierarchyFeature } from './callHierarchy'
 import { SemanticTokensFeature } from './semanticTokens'
 import { InlayHintsFeature } from './inlayHint'
+import { WorkspaceSymbolFeature } from './workspaceSymbol'
 import { LinkedEditingFeature } from './linkedEditingRange'
 import { DidCreateFilesFeature, DidDeleteFilesFeature, DidRenameFilesFeature, WillCreateFilesFeature, WillDeleteFilesFeature, WillRenameFilesFeature } from './fileOperations'
 
@@ -582,6 +583,9 @@ export class LanguageClient extends BaseLanguageClient {
     }
     if (!disabledFeatures.includes('inlayHint')) {
       this.registerFeature(new InlayHintsFeature(this))
+    }
+    if (!disabledFeatures.includes('workspaceSymbol')) {
+      this.registerFeature(new WorkspaceSymbolFeature(this))
     }
     if (!disabledFeatures.includes('workspaceFolders')) {
       this.registerFeature(new WorkspaceFoldersFeature(this))
